Allow tor-rtcompat to build with no runtimes.
Thanks to cargo's version-2 feature resolver, we can require a
runtime for tests only.

I'm also making it so that the functions that create or fetch
Runtimes only exist when one of the runtime features is enabled.
For now that seems like a better solution than having those
functions exist but panic.

Closes #129.