tor-circmgr: Improve coverage on build.rs.
The good news is that this tests the succeeding and failing cases,
and even managed to expose an off-by-one error in our hop counting.

The bad news is that these tests have shown even more ways in which
our wait_for code is unreliable, and forced me to slow it down even
harder.