Refactor double_timeout() to make its timeouts futures in advance
Since these futures rely on the current time when they decide when
to expire, we don't want them looking at the current time from
inside the call to runtime.spawn or after that call.  Instead we'd
like them to look immediately on construction.

This approach is based on a comment from janimo on #149.  It doesn't
fix #149, though.