Upgrade to latest fslock and use its per-fd exclusion feature.
The latest version of the fslock crate has an optional feature that
allows file-locks to exclude one another even if they are opened
within the same process.  On Windows this behavior comes for free;
on Unix it depends on an internal map of (dev,inode) pairs.

Here we upgrade to the latest version of the fslock crate, and use
this feature.