Basic implementation for prop#332 ntor v3 backend.
This isn't tied into the protocol code yet, since we don't actually
have anybody to negotiate with here.  We'll need some API tweaks in
order to express the sending of extended information too.  For now,
I'm putting it behind a feature, but if we build this into the tor
network, the feature-gate should go away.