CI: Fix async-std task
Previously it didn't enable any TLS provider, since we made
native-tls optional a few commits ago.  Now it enables rustls,
so that rustls also gets a quick check along with async-std.

I've also switched this test to use "cargo clippy" in place of
"cargo test" because it's a strict superset.